async func rewrite; global const added
