+++ src/constants/constants.js
@@ -1,0 +1,3 @@
+const API_URL = 'https://64d8eebd5f9bf5b879ceb6cd.mockapi.io/products';
+
+export default API_URL;
